+++ off-chain/src/core/auction-storage-interface.js
