+++ off-chain/test/use-cases/auction-manager.test.js
@@ -1,6 +1,6 @@
 /* eslint-disable no-unused-vars */
 import AuctionManager from '../../src/core/auction-manager.js';
-import MemAuctionStorage from '../../src/infra/memAuctionStorage.js';
+import MemAuctionStorage from '../../src/infra/mem-auction-storage.js';
 import ValidatorMock from '../validator_mock.js';
 
 const validatorMock = new ValidatorMock();
